+++ src/shared-types.ts
@@ -1,1 +1,10 @@
-export type WcsSize = 's' | 'm' | 'l' | 'xl';
+/**
+ * Meant to be extracted with TS and used in each component differently, depending on the use case.
+ * Global size guideline (with basis at 8px) :
+ * xl = 56 px
+ * l = 48 px
+ * m = 40 px (default)
+ * s = 32 px
+ * xs = 24 px
+ */
+export type WcsSize = 'xs' | 's' | 'm' | 'l' | 'xl';