+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - **button**: add `target` (`"_blank" | "_self"`, null by default) to choose where to open links when using `href`
+- **button**: add the following sizes : `l`, `m (default)`, `s`
 - **doc**: add documentation on how to initialize an application with the business or communication design systems
 - **ci**: add the ability to deploy any branch on a dedicated environment
 
@@ -24,6 +25,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - **styles**: Set WCS global variable "base size" to 8px. Has no impact to previous styles.
 - **types**: Create and set index as root file for types export (generated & custom types)
 - **types**: Create a new shared-types file for custom types, containing WcsSize
+- **button**: add the size attributes with available values `l`, `m (default)`, `s`
 
 ### Deprecated
 
@@ -31,7 +33,7 @@ for soon-to-be removed features.
 
 ### Removed
 
-for now removed features.
+- **button**: remove the shape="small" attribute. (Replace it with size="s").
 
 ### Fixed